chore: rename event.ts to events.ts
